+++ RELEASE_NOTES.md
@@ -1,4 +1,6 @@
 Version 1.5.0
 ===============
 - Added an intent for listing a user's loadouts that are currently persisted.
+- Fixing an issue with the popular Trials weapons returning an empty string if
+    the user hasn't played any matches.